Moved PackedBubble integration.
